% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{as.vbscript_lines}
\alias{as.vbscript_lines}
\alias{as.vbscript_lines.character}
\title{\code{vbscript_lines}}
\usage{
as.vbscript_lines(x)

\method{as.vbscript_lines}{character}(x)
}
\arguments{
\item{x}{\code{[character]} (mandatory, no default)

character string vector to coerce to \code{vbscript_lines}}
}
\description{
Coercion to class \code{vbscript_lines}
}
\section{Methods (by class)}{
\itemize{
\item \code{character}: coerces character string vector to class
`vbscript_lines``
}}

