% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{print.vbscript_lines}
\alias{print.vbscript_lines}
\title{Print \code{vbscript_lines}}
\usage{
\method{print}{vbscript_lines}(x, max.print = 50, ...)
}
\arguments{
\item{x}{\code{[vbscript_lines]} (mandatory, no default)

a \code{vbscript_lines} object}

\item{max.print}{\code{[integer, numeric]} (mandatory, default 50)

maximum number of lines allowed to be printed; if \code{x} has more elements
than this, only the fist 10 and last 10 elements are shown in print}

\item{...}{added for compatibility with \link{print}}
}
\description{
Print method for \code{vbscript_lines} objects
}
