% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{vbscript_code_generation_utils}
\alias{vbscript_code_generation_utils}
\alias{vbscript_lines_echo}
\alias{vbscript_lines_set_focus_to_window}
\alias{vbscript_lines_execute_keystrokes}
\title{\code{vbscript} Code Generation}
\usage{
vbscript_lines_echo(x)

vbscript_lines_set_focus_to_window(window_name)

vbscript_lines_execute_keystrokes(keystrokes, init_shell = TRUE)
}
\arguments{
\item{x}{\code{[character]} (mandatory, no default)

a string to echo in vbscript}

\item{window_name}{\code{[character]} (mandatory, no default)

set Windows focus to window by this name; if such a window does not exist,
this function just fails to set focus to anything
(without indicating failure in R or otherwise)}

\item{keystrokes}{\code{[character]} (mandatory, no default)

one or more keystrokes to execute; specials include \code{{ENTER}} for Enter,
\code{{TAB}} for Tab, etc.}

\item{init_shell}{\code{[logical]} (mandatory, default \code{TRUE})

keystrokes can only be executed if \code{WScript.Shell} has been initiated earlier
in the script; if this argument is \code{TRUE}, this initiation is preprended to
output of this function}
}
\value{
a \code{vbscript_lines} object (see \link{as.vbscript_lines})
}
\description{
Utilities to generate vbscript code in R.
}
\details{
\itemize{
\item \code{vbscript_lines_echo}: simply generates a \code{Wscript.Echo} call using \code{x}
}

\itemize{
\item \code{vbscript_lines_set_focus_to_window}: generates a \code{FocusShell.AppActivate}
call using \code{window_name};  if such a window does not exist,
this function just fails to set focus to anything
(without indicating failure in R or otherwise)
}

\itemize{
\item \code{vbscript_lines_execute_keystrokes}: generates \code{WshShell.SendKey}
calls using \code{keystrokes}
}
}
