% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{vbscript_file_utils}
\alias{vbscript_file_utils}
\alias{can_call_vbscript_files}
\alias{write_vbscript_file}
\alias{read_vbscript_file}
\alias{call_vbscript_file}
\alias{call_vbscript_lines}
\title{vbscript Files}
\usage{
can_call_vbscript_files()

write_vbscript_file(x, file_path, ...)

read_vbscript_file(file_path, ...)

call_vbscript_file(file_path)

call_vbscript_lines(x)
}
\arguments{
\item{x}{\code{[character]} (mandatory, no default)
\itemize{
\item \code{write_vbscript_file}: lines to write to file
\item \code{call_vbscript_lines}: lines to call
}}

\item{file_path}{\code{[character]} (mandatory, no default)

path where to write file / read file from}

\item{...}{\itemize{
\item \code{write_vbscript_file}: arguments passed to \link{writeLines},
except \code{text} and \code{con}
\item \code{read_vbscript_file}: arguments passed to \link{readLines}, expect \code{con}
}}
}
\description{
Utilities to read, write, and call vbscript files.
}
\details{
\itemize{
\item \code{can_call_vbsript_files}: returns \code{TRUE} if you can call vbscript files on
this system; this function tries calling \link{call_vbscript_lines} with a very
simple command
}

\itemize{
\item \code{write_vbscript_file}: simply calls \link{writeLines} after some checks
}

\itemize{
\item \code{read_vbscript_file}: simply calls \link{readLines} after some checks
and returns a \code{vbscript_lines} object (see \link{as.vbscript_lines})
}

\itemize{
\item \code{call_vbscript_file}: calls the specified file as a vbscript file using
Windows executable \code{cscript.exe}; output and errors are captured into R
and returned; see \link{system2}
}

\itemize{
\item \code{call_vbscript_lines}: writes \code{x} into a temporary file and passes that
file's path to \link{call_vbscript_file}
}
}
